refactor: apply DRY principle and move the card display to the separated partials
